requirements_render.txt
Flask==2.3.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Authlib==1.2.1
reportlab==4.0.4
gunicorn==21.2.0
Werkzeug==2.3.7
requests==2.31.0
# numpy, pandas, scikit-learn are typically pre-installed or installed separately
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
